feat: 매너 온도 반영 기능 구현
- 매너 온도 업데이트
- 베스트 플레이어 등록

+++ linkup/src/main/java/com/learningcrew/linkup/common/dto/query/MeetingMemberDto.java
@@ -14,6 +14,6 @@ public class MeetingMemberDto {
     private String nickname;                  // 닉네임
     private LocalDate birthDate;              // 생년월일
     private String introduction;              // 자기소개
-    private BigDecimal mannerTemperature;     // 매너 온도
+    private Double mannerTemperature;     // 매너 온도
     private String profileImageUrl;           // 프로필 이미지
 }
